pyproject.toml
[project]
name = "bbs-mainframe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "pycryptodome>=3.21.0",
    "sqlalchemy>=2.0.36",
]

[dependency-groups]
dev = [
    "black>=24.10.0",
    "isort>=5.13.2",
]
